Splitting POIAnalyticsControlPanel into components
